+++ lib/pages/register/registerPrestador/registerPagePrestador2.dart
@@ -1,16 +1,18 @@
 import 'package:flutter/material.dart';
+import 'package:motohelp/components/button_elevated.dart';
 
-class RegisterPageMotoka2 extends StatefulWidget {
-  const RegisterPageMotoka2({super.key});
+class RegisterPagePrestador2 extends StatefulWidget {
+  const RegisterPagePrestador2({super.key});
 
   @override
-  State<RegisterPageMotoka2> createState() => _RegisterPageMotoka2State();
+  State<RegisterPagePrestador2> createState() => _RegisterPagePrestador2State();
 }
 
-class _RegisterPageMotoka2State extends State<RegisterPageMotoka2> {
+class _RegisterPagePrestador2State extends State<RegisterPagePrestador2> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: AppBar(),
       body: Padding(
         padding: const EdgeInsets.all(30.0),
         child: Column(
@@ -64,35 +66,6 @@ class _RegisterPageMotoka2State extends State<RegisterPageMotoka2> {
                           Radius.circular(10),
                         ),
                       ),
-                      hintText: "CRLV",
-                      hintStyle: TextStyle(
-                        color: Colors.black,
-                        fontFamily: 'tahoma',
-                        fontWeight: FontWeight.bold,
-                      ),
-                      fillColor: Colors.blue,
-                      filled: true,
-                    ),
-                  ),
-                  SizedBox(
-                    height: 15,
-                  ),
-                  TextFormField(
-                    textAlign: TextAlign.center,
-                    decoration: InputDecoration(
-                      prefixIcon: Padding(
-                        padding: EdgeInsets.all(5),
-                        child: Icon(
-                          Icons.camera_alt_outlined,
-                          color: Colors.black,
-                        ),
-                      ),
-                      contentPadding: EdgeInsets.all(10),
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.all(
-                          Radius.circular(10),
-                        ),
-                      ),
                       hintText: "Foto do rosto",
                       hintStyle: TextStyle(
                         color: Colors.black,
@@ -109,20 +82,7 @@ class _RegisterPageMotoka2State extends State<RegisterPageMotoka2> {
                 ],
               ),
             ),
-            Padding(
-              padding: EdgeInsets.all(15),
-              child: Container(
-                child: ElevatedButton(
-                  onPressed: () {},
-                  child: Text(
-                    'Finalizar',
-                    style: TextStyle(
-                      color: Colors.black,
-                    ),
-                  ),
-                ),
-              ),
-            ),
+            ButtonElevated(textButton: 'Finalizar', navigatorIndicator: '/new'),
           ],
         ),
       ),
